User Endpoint Implemented

+++ pujo/views.py
@@ -6,13 +6,26 @@ from .models import Pujo
 from .serializers import PujoSerializer, TrendingPujoSerializer
 from core.ResponseStatus import ResponseStatus
 import logging
+from django.utils import timezone
+from user.permission import IsSuperOrAdminUser
+from rest_framework_simplejwt.authentication import JWTAuthentication
+from rest_framework.decorators import action
+from rest_framework import permissions
 
 logger = logging.getLogger("pujo")
 
 class PujoViewSet(viewsets.ModelViewSet):
     queryset = Pujo.objects.all()
     serializer_class = PujoSerializer
-    lookup_field = 'uuid'
+    lookup_field = 'id'
+    permission_classes=[IsSuperOrAdminUser]
+    authentication_classes = [JWTAuthentication]
+
+    def get_permissions(self):
+        if self.action in ['list', 'retrieve', 'trending']:
+            # Allow anyone to see list,trending and retreive
+            return [permissions.AllowAny()]
+        return super().get_permissions()
 
     def list(self, request, *args, **kwargs):
         try:
@@ -94,9 +107,9 @@ class PujoViewSet(viewsets.ModelViewSet):
             }
             return Response(response_data, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
-    def retrieve(self, request, pk=None, *args, **kwargs):
+    def retrieve(self, request, uuid=None, *args, **kwargs):
         try:
-            pujo = self.get_object()
+            pujo = self.get_queryset().filter(id=uuid).first()
             serializer = self.get_serializer(pujo)
             response_data = {
                 'result': serializer.data,
@@ -111,11 +124,14 @@ class PujoViewSet(viewsets.ModelViewSet):
             return Response(response_data, status=status.HTTP_404_NOT_FOUND)
 
     def create(self, request, *args, **kwargs):
+        user = request.user
+        self.check_object_permissions(request, user)
+        print("Current user:", request.user)
         serializer = self.get_serializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
             response_data = {
-                'result': {'id': serializer.data["uuid"]},
+                'result': {'id': serializer.data["id"]},
                 'status': ResponseStatus.SUCCESS.value
             }
             return Response(response_data, status=status.HTTP_201_CREATED)
@@ -126,12 +142,14 @@ class PujoViewSet(viewsets.ModelViewSet):
             }
             return Response(response_data, status=status.HTTP_400_BAD_REQUEST)
 
-    def update(self, request, pk=None, *args, **kwargs):
+    def update(self, request, uuid=None, *args, **kwargs):
         try:
-            pujo = self.get_object()
+            user = request.user
+            self.check_object_permissions(request, user)
+            pujo = self.get_queryset().filter(id=uuid).first()
             serializer = self.get_serializer(pujo, data=request.data)
             if serializer.is_valid():
-                serializer.save()
+                serializer.save(updated_at=timezone.now())
                 response_data = {
                     'result': serializer.data,
                     'status': ResponseStatus.SUCCESS.value
@@ -150,9 +168,11 @@ class PujoViewSet(viewsets.ModelViewSet):
             }
             return Response(response_data, status=status.HTTP_404_NOT_FOUND)
 
-    def destroy(self, request, pk=None, *args, **kwargs):
+    def destroy(self, request, uuid=None, *args, **kwargs):
         try:
-            pujo = self.get_object()
+            user = request.user
+            self.check_object_permissions(request, user)
+            pujo = self.get_queryset().filter(id=uuid).first()
             pujo.delete()
             response_data = {
                 'result': "Delete successful",
